Moved routes and navigationOptions into a separate file in the config module
