+++ app/components/navigationComponents/BackButtonHeader/assets/ic_back.png
