+++ app/components/navigationComponents/ImageHeader/styles.js
